pyproject.toml
[tool.poetry]
name = "dodona"
version = "0.1.0"
description = "Simple GUI for the OpenAI API, made with OpenAI"
authors = ["Iwan van der Kleijn <iwanvanderkleijn@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.5.0"
openai = "^0.27.7"
sounddevice = "^0.4.6"
soundfile = "^0.12.1"
numpy = "^1.24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
